+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.2.0
+- Sockets now have "kind" attribute
+- TelnetSocket.getOption() now available
+- TelnetServer now requires local/remote options
+
 ## 2.1.0
 - Added WebServer, print, println, prompt; all untested